rename bytes_chunk_count
