+++ ext/bpeboost/bytes_chunk_count.pyi
